+++ packages/rma-server/src/constants/agenda-job.ts
@@ -13,3 +13,28 @@ export const AGENDA_JOB_METADATA = {
   lastRunAt: new Date(),
   lastFinishedAt: new Date(),
 };
+
+export function getParsedPostingDate(payload) {
+  let date: Date;
+  try {
+    date = new Date(
+      `${payload.posting_date} ${payload.posting_time || '00:00:00'}`,
+    );
+  } catch {}
+  if (date && isNaN(date?.getMilliseconds())) {
+    date = new Date(
+      `${this.parsePostingDate(payload.posting_date)} ${
+        payload.posting_time || '00:00:00'
+      }`,
+    );
+  }
+  if (date && isNaN(date?.getMilliseconds())) {
+    date = new Date();
+  }
+  return date;
+}
+
+export function parsePostingDate(posting_date) {
+  const splitDate = posting_date.split('-');
+  return `${splitDate[1]}-${splitDate[0]}-${splitDate[2]}`;
+}