+++ packages/rma-server/src/constants/app-strings.ts
@@ -123,6 +123,7 @@ export const DELIVERY_NOTE_DOCTYPE = 'Delivery Note';
 export const STOCK_ENTRY_SERIALS_BATCH_SIZE = 6000;
 export const DELIVERY_NOTE_SERIAL_BATCH_SIZE = 6000;
 export const SERIAL_NO_VALIDATION_BATCH_SIZE = 10000;
+export const ITEM_SYNC_BUFFER_COUNT = 30;
 export const FRAPPE_INSERT_MANY_BATCH_COUNT = 2;
 export const PURCHASE_RECEIPT_DOCTYPE_NAME = 'Purchase Receipt';
 export const SERIAL_NO_DOCTYPE_NAME = 'Serial No';