+++ packages/rma-server/src/print/aggregates/print-aggregate/print-aggregate.service.ts
@@ -20,7 +20,7 @@ export class PrintAggregateService {
 
     res.set({
       'Content-Type': 'application/pdf',
-      'Content-Disposition': `attachment; filename=${invoice.name}.pdf`,
+      'Content-Disposition': `attachment; filename=${invoice.print.print_type}.pdf`,
       'Content-Length': buffer.length,
     });
 
@@ -80,7 +80,8 @@ export class PrintAggregateService {
       .fontSize(10)
       .text(invoice.name, cord.x, cord.y, { align: 'left' })
       .text(invoice.territory, cord.x, cord.y, { align: 'right' });
-    this.generateHr(doc, 155);
+    doc.moveDown();
+    this.generateHr(doc, doc.y);
 
     const customerInformationTop = doc.y;
 
@@ -104,10 +105,34 @@ export class PrintAggregateService {
       .text('Created By:', 300, customerInformationTop + 45)
       .text(invoice.created_by, 420, customerInformationTop + 45)
       .text('Approved By:', 300, customerInformationTop + 60)
-      .text(invoice.modified_by, 420, customerInformationTop + 60)
-      .moveDown()
-      .moveDown()
-      .moveDown();
+      .text(invoice.modified_by, 420, customerInformationTop + 60);
+
+    if (invoice?.print?.s_warehouse || invoice?.print?.t_warehouse) {
+      if (invoice?.print?.s_warehouse) {
+        doc
+          .fontSize(11)
+          .fillColor('#000000')
+          .text(
+            `From Warehouse: ${invoice.print.s_warehouse}`,
+            50,
+            customerInformationTop + 95,
+          );
+      }
+
+      if (invoice?.print?.t_warehouse) {
+        doc
+          .fontSize(11)
+          .fillColor('#000000')
+          .text(
+            `To Warehouse: ${invoice.print.t_warehouse}`,
+            300,
+            customerInformationTop + 95,
+            { align: 'right' },
+          );
+      }
+    }
+
+    doc.moveDown().moveDown();
   }
 
   generatePrintTable(doc, invoice: DeliveryChalanDto) {
@@ -144,7 +169,13 @@ export class PrintAggregateService {
     }
     doc.moveDown();
 
-    this.generateTableRow(doc, doc.y, '', { name: 'Total' }, invoice.total_qty);
+    this.generateTableRow(
+      doc,
+      doc.y,
+      '',
+      { name: 'Total' },
+      invoice.total_qty || this.getItemTotal(invoice.items),
+    );
     doc.moveDown();
     doc.moveDown();
     doc.moveDown();
@@ -176,12 +207,20 @@ export class PrintAggregateService {
   ) {
     doc.moveDown();
     const height = doc.y;
-    doc.fontSize(10).text(id, 50, height);
+    doc.fontSize(10).fillColor('#000000').text(id, 50, height);
     doc.text(item.name, 100, height, { width: 390 });
     doc.text(quantity, 450, height, { align: 'right' });
     if (item.serials) {
-      doc.text(this.getSerialKeys(item), 100, doc.y, { width: 390 });
+      doc
+        .fillColor('#444444')
+        .text(this.getSerialKeys(item), 100, doc.y, { width: 390 });
     }
+  }
+
+  getItemTotal(items: any[]) {
+    let total = 0;
+    items.forEach(item => (total += item.qty));
+    return total;
   }
 
   getSerialKeys(item: { name: string; serials?: string }) {